refactor: move health diagnostics contracts to output crate
